More consistent type Long to Int

+++ src/main/scala/org/edla/tmdb/api/Protocol.scala
@@ -7,27 +7,27 @@ object Protocol extends DefaultJsonProtocol {
   case class AuthenticateResult(expires_at: String, request_token: String, success: Boolean)
   case class ProductionCountry(iso_3166_1: String, name: String)
   case class Release(iso_3166_1: String, certification: String, release_date: String)
-  case class Genre(id: Long, name: String)
-  case class ProductionCompanie(name: String, id: Long)
+  case class Genre(id: Int, name: String)
+  case class ProductionCompanie(name: String, id: Int)
   case class SpokenLanguage(iso_639_1: String, name: String)
-  case class Collection(poster_path: Option[String], id: Long, name: String, backdrop_path: Option[String])
-  case class Movie(runtime: Option[Long],
+  case class Collection(poster_path: Option[String], id: Int, name: String, backdrop_path: Option[String])
+  case class Movie(runtime: Option[Int],
                    status: String,
                    backdrop_path: Option[String],
                    overview: Option[String],
                    title: String,
-                   vote_count: Long,
+                   vote_count: Int,
                    tagline: Option[String],
                    belongs_to_collection: Option[Collection],
                    original_title: String,
                    poster_path: Option[String],
                    production_countries: List[ProductionCountry],
-                   revenue: Long,
+                   revenue: Int,
                    homepage: Option[String],
                    imdb_id: String,
-                   id: Long,
+                   id: Int,
                    release_date: Option[String],
-                   budget: Long,
+                   budget: Int,
                    popularity: Double,
                    genres: List[Genre],
                    production_companies: List[ProductionCompanie],
@@ -36,12 +36,12 @@ object Protocol extends DefaultJsonProtocol {
   case class Result(original_title: String,
                     poster_path: Option[String],
                     release_date: Option[String],
-                    id: Long,
+                    id: Int,
                     adult: Boolean,
                     title: String,
                     popularity: Double,
-                    vote_count: Long,
-                    vote_average: Long,
+                    vote_count: Int,
+                    vote_average: Int,
                     backdrop_path: Option[String])
   case class Results(total_results: Int, results: List[Result], page: Int, total_pages: Int)
   case class Images(still_sizes: List[String],
@@ -52,22 +52,22 @@ object Protocol extends DefaultJsonProtocol {
                     logo_sizes: List[String],
                     backdrop_sizes: List[String])
   case class Configuration(images: Images, change_keys: List[String])
-  case class Cast(cast_id: Long,
+  case class Cast(cast_id: Int,
                   character: String,
                   credit_id: String,
                   name: String,
-                  order: Long,
+                  order: Int,
                   profile_path: Option[String])
   case class Crew(credit_id: String,
                   departement: Option[String],
-                  id: Long,
+                  id: Int,
                   job: String,
                   name: String,
                   profile_path: Option[String])
-  case class Credits(id: Long, cast: List[Cast], crew: List[Crew])
-  case class Releases(id: Long, countries: List[Release])
+  case class Credits(id: Int, cast: List[Cast], crew: List[Crew])
+  case class Releases(id: Int, countries: List[Release])
 
-  case class Error(status_code: Long, status_message: String)
+  case class Error(status_code: Int, status_message: String)
 
   val noCrew     = Crew("", None, 0, "", "Unknown", None)
   val unReleased = Release("", "", "Unknown")
